standard-backend/src/requirements.txt
boto3
flask-lambda-python36
flask-swagger-ui
Flask-WTF
flask==2.1.3
matplotlib
memory-profiler
numpy
opencv-python
Pillow
Werkzeug==2.0.3
